VMS-7368: Deleted unused maven-built distros cloud_db-deb and connection_mediator-deb
VMS-8241: Refactored: files that are copied to metadata_sdk.zip moved to "metadata_sdk" subdir
VMS-8113: For detected objects use colors from a predefined set
VMS-7981: Make analytics zoom windows aspect ratio the same as the source video aspect ratio
VMS-8113: Do not display empty object tooltips
VMS-8024 - Camera discovery functions moved to rest connection
VMS-8024 - Renamed and relocated files
VMS-8024 - Small review fixes
VMS-7916: review fixes
VMS-7916: review fixes

<==== Mercurial data ====>
Branch: mobile

+++ plugins/metadata/stub_metadata_plugin/src/nx/mediaserver_plugins/metadata/stub/camera_manager.cpp
@@ -183,7 +183,7 @@ MetadataPacket* CameraManager::cookSomeObjects()
     *reinterpret_cast<int*>(objectId.bytes) = static_cast<int>(intPart);
     commonObject->setId(objectId);
 
-    commonObject->setBoundingBox(Rect(dt, dt, 0.25, 0.25));
+    commonObject->setBoundingBox(Rect((float) dt, (float) dt, 0.25F, 0.25F));
 
     auto objectPacket = new CommonObjectsMetadataPacket();
     objectPacket->setTimestampUsec(m_videoFrame->timestampUsec());
